fix: function ft_strlen

+++ libft.h
@@ -1,5 +1,8 @@
+#include <unistd.h>
+#include <stdlib.h>
+
 int	ft_isalpha(int c);
 int	ft_isdigit(int c);
 int	ft_isalnum(int c);
 int	ft_isprint(int c);
-int	ft_strlen(const char *str);
+size_t	ft_strlen(const char *str);
